Created some tests for the job queue

+++ master/utils/cleaner.py
@@ -4,9 +4,9 @@ from threading import Thread
 
 
 class Cleaner(ABC):
-    def __init__(self, interval: int):
+    def __init__(self, interval: float):
         self._cleaning_thread = Thread(target=self._cleaning_loop, daemon=True)
-        self._interval = interval // 1000
+        self._interval = interval
         self._cleaning_thread.start()
 
     def _cleaning_loop(self) -> None:
